Refactor: removed assign_operator_extension in favor of lexer
While moving most but the actual expression handling from the lexer to
extension traits and moving the purpose of the lexer from operator only
to handle the complete syntax operator_precedence_parser is renamed to
lever.

The tests are split up into the matching extensions instead of the lexer
itself.

The operation enum is moved to operations.rs.
